separate toInput extensions again
