+++ common/src/main/scala/explore/common/ConstraintsQueries.scala
@@ -20,8 +20,7 @@ import lucuma.schemas.ObservationDB.Types.*
 import monocle.Lens
 import org.typelevel.log4cats.Logger
 import queries.common.ObsQueriesGQL.*
-import queries.schemas.odb.conversions.*
-
+import queries.schemas.odb.ODBConversions.*
 object ConstraintsQueries:
   case class UndoView(
     obsIds:  List[Observation.Id],
