+++ explore/src/main/scala/explore/tabs/ConfigurationTile.scala
@@ -18,7 +18,7 @@ import lucuma.core.model.Observation
 import lucuma.ui.syntax.all.*
 import lucuma.ui.syntax.all.given
 import org.typelevel.log4cats.Logger
-import queries.schemas.itc.conversions.*
+import queries.schemas.itc.ITCConversions.*
 import queries.schemas.odb.ObsQueries.*
 
 object ConfigurationTile {
